feat: Deletion and redeployment
