+++ src/components/ui/buttonVariants.tsx
@@ -38,6 +38,8 @@ export const buttonVariants = cva(
 				positive: 'bg-green text-white shadow-xs hover:bg-green/90',
 				warning: 'bg-yellow text-white shadow-xs hover:bg-yellow/90',
 				outline: `${outlineCommon} ${hoverBounce}`,
+				ghostOutline:
+					`${outlineCommon} ${hoverBounce} border-none`,
 				positiveOutline:
 					`${outlineCommon} ${hoverBounce} border-green`,
 				destructiveOutline:
