+++ src/features/instance/applications/components/ApplicationsSidebar/index.tsx
@@ -38,11 +38,9 @@ export function ApplicationsSidebar() {
 		}
 	}, [openedEntry, focusedItem, items]);
 
-	// TODO: application command with shortcuts for a new file and folder
-	// TODO: delete file
-	// TODO: delete folder
-	// TODO: redeploy package
-	// TODO: restart cluster
+	// TODO: application command with shortcuts for creation and deletion
+	// TODO: consistently drive file and folder selection through context (particularly during creation and deletion)
+	// TODO: Split all this logic up into smaller files, right?
 	// TODO: onRenameItem f2 handling
 	// TODO: on drag item from one folder to another
