+++ src/features/instance/applications/modals/AddDirectoryOrFileModal.tsx
@@ -108,6 +108,8 @@ export function AddDirectoryOrFileModal({
 										<Input
 											disabled={isPending}
 											type="text"
+											autoComplete="off"
+											autoCapitalize="off"
 											{...field}
 										/>
 									</FormControl>
